+++ src/ui/popup.rs
@@ -1,5 +1,5 @@
 use crate::app::App;
-use crate::ui::ui::surrounding_block;
+use crate::ui::surrounding_block;
 use tui::{
     backend::Backend,
     layout::{Constraint, Direction, Layout, Rect},
