+++ src/ui/view.rs
@@ -1,6 +1,6 @@
 use crate::{
     app::{InputMode, MainMode, NavMode, SearchMode, View},
-    ui::{ui::surrounding_block, utils::htmlparser},
+    ui::{surrounding_block, utils::htmlparser},
 };
 use dot_graph::Node;
 use rayon::prelude::*;
@@ -15,12 +15,7 @@ use tui::{
     Frame,
 };
 
-pub fn draw_view<B: Backend>(
-    f: &mut Frame<B>,
-    chunk: Rect,
-    mmode: &MainMode,
-    view: &mut View,
-) {
+pub fn draw_view<B: Backend>(f: &mut Frame<B>, chunk: Rect, mmode: &MainMode, view: &mut View) {
     let chunks = Layout::default()
         .direction(Direction::Horizontal)
         .constraints([Constraint::Percentage(35), Constraint::Percentage(65)].as_ref())
